servidor beta 2
proyecto con la segunda beta del servidor , esta versión unifico las dos clases del servidor en una sola
