+++ MasterClientApps/src/controller/Main.java
@@ -21,16 +21,17 @@ import modelo.LectorUsuarios;
  */
 public class Main extends Application {
 
-    /**
-     * se declara el contructor del metodo main para poder hacer ejecutable el
-     * proyecto
-     *
-     * @param args se envian los argumentos y comandos de linea
-     */
+
+
+        /**
+         * se declara el contructor del metodo main para poder hacer ejecutable el
+         * proyecto
+         *
+         * @param args se envian los argumentos y comandos de linea
+         */
     public static void main(String[] args) {
         launch(args);
     }
-
     @Override
     /**
      * Método principal que genera el hilo necesario para poder ejecutar la GUI
@@ -63,6 +64,8 @@ public class Main extends Application {
         stageMaster.show();
         stageClient.show();
 
+
     }
+
 
 }
